Move AbstractAuthMethod
